Removing a warning/error.  The sprintf function when %lu is specified expects a "long unsigned int" but the argument supplied was a uint32_t {aka unsugned int}. In this case the argument actually is a long unsigned int so added a cast so the argument would be seen without a warning.  Minor nit but best to remove warnings so that it's easy to see when a real issue happens.

+++ Code/Adafruit_FONA.cpp
@@ -1883,7 +1883,7 @@ boolean Adafruit_FONA::postData(const char *request_type, const char *URL, const
 
     char dataBuff[sizeof(bodylen) + 20];
 
-    sprintf(dataBuff, "AT+HTTPDATA=%lu,10000", bodylen);
+    sprintf(dataBuff, "AT+HTTPDATA=%lu,10000", (long unsigned int)bodylen);
     if (! sendCheckReply(dataBuff, "DOWNLOAD", 10000))
       return false;
